requirements.txt
aiologger==0.7.0
alembic==1.13.1
annotated-types==0.5.0
anyio==4.4.0
asyncio==3.4.3
asyncpg==0.28.0
attrs==24.1.0
bcrypt==4.0.1
black==23.11.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
colorlog==6.8.2
cryptography==41.0.3
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.1
greenlet==2.0.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.0
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.3
loguru==0.7.2
Mako==1.3.3
MarkupSafe==2.1.5
mypy-extensions==1.0.0
outcome==1.3.0.post0
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==4.1.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic-settings==2.0.3
pydantic_core==2.4.0
pydentic==0.0.1.dev3
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.9
python-stdnum==1.18
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
starlette==0.27.0
tomli==2.0.1
trio==0.26.0
typing_extensions==4.7.1
uvicorn==0.23.1
uvloop==0.19.0
watchfiles==0.19.0
webencodings==0.5.1
websockets==11.0.3
